+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/activity/history/ListingActivityUtil.java
@@ -15,7 +15,6 @@ import gov.healthit.chpl.domain.CertificationStatusEvent;
 import gov.healthit.chpl.domain.CertifiedProductSearchDetails;
 import gov.healthit.chpl.exception.EntityRetrievalException;
 import gov.healthit.chpl.service.CertificationCriterionService;
-import gov.healthit.chpl.util.Util;
 import lombok.extern.log4j.Log4j2;
 
 @Log4j2
@@ -59,13 +58,15 @@ public class ListingActivityUtil {
     }
 
     private void ensureCertificationResultHasCriterion(CertificationResult result) {
-        //This method assumes that if the Criterion does not exist, the criterion is not the Cures version
+        //This method assumes that if the Criterion does not exist, the criterion is not the Cures version.
+        //Once we added Cures criteria into the system we also started filling in the certResult.criterion value
+        //so that's why we assume if there's a null certResult.criterion value the thing to fill in must be
+        //the non-cures criterion.
         if (result.getCriterion() == null) {
             List<CertificationCriterion> criteria = certificationCriterionService.getByNumber(result.getNumber());
-            //Get the first non-Cures criterion
+            //Get the criterion with the earliest active start date and assume this is the non-cures one
             result.setCriterion(criteria.stream()
-                    .filter(c -> !Util.isCures(c))
-                    .findFirst()
+                    .min((o1, o2) -> o1.getStartDay().compareTo(o2.getStartDay()))
                     .get());
 
         }